+++ src/App.tsx
@@ -1,14 +1,15 @@
 import React from "react";
-import Home from "./pages/home";
+import HomePage from "./pages/home/HomePage";
 import { Routes, Route } from "react-router-dom";
 import PrivateRoute from "./utils/router/privateRoute";
-import AuthRootComponent from "./pages/auth";
+import AuthRootComponent from "./pages/auth/AuthRootComponent";
 import { ColorModeContext, useMode } from "./theme";
 import { CssBaseline, ThemeProvider } from "@mui/material";
 import LayoutComponent from "./components/layout/LayoutComponent";
-import WatchlistComponent from "./pages/watchlist";
-import NewsComponent from "./pages/news";
-import SettingsComponent from "./pages/settings";
+import WatchlistPage from "./pages/watchlist/WatchlistPage";
+import NewsPage from "./pages/news/NewsPage";
+import SettingsPage from "./pages/settings/SettingsPage";
+import { SingleAssetPage } from "./pages/single-asset";
 
 function App() {
   const [theme, colorMode] = useMode();
@@ -21,10 +22,11 @@ function App() {
           <Routes>
             <Route element={<LayoutComponent />}>
               <Route element={<PrivateRoute />}>
-                <Route path="/" element={<Home />} />
-                <Route path="/watchlist" element={<WatchlistComponent />} />
-                <Route path="/news" element={<NewsComponent />} />
-                <Route path="/settings" element={<SettingsComponent />} />
+                <Route path="/" element={<HomePage />} />
+                <Route path="/watchlist" element={<WatchlistPage />} />
+                <Route path="/news" element={<NewsPage />} />
+                <Route path="/settings" element={<SettingsPage />} />
+                <Route path="/single/:id"  element={<SingleAssetPage />} />
               </Route>
               <Route path="login" element={<AuthRootComponent />} />
               <Route path="register" element={<AuthRootComponent />} />
