+++ src/components/charts/area-chart/index.ts
@@ -1,0 +1,1 @@
+export { AreaChartComponent } from "./AreaChartComponent";
