+++ src/components/charts/line-chart/index.ts
@@ -1,0 +1,1 @@
+export { LineChartComponent } from "./LineChartComponent";
