+++ src/pages/auth/index.ts
@@ -1,0 +1,3 @@
+export { AuthRootComponent } from "./AuthRootComponent";
+
+
