+++ src/pages/auth/login/index.ts
@@ -1,0 +1,3 @@
+export { LoginPage } from "./LoginPage";
+
+
