+++ src/pages/auth/login/style.module.css
@@ -13,4 +13,5 @@
  }
  .form{
    text-align: center;
+   font-size: 32;
  }
