+++ src/pages/auth/style.module.css
@@ -4,7 +4,7 @@
   align-items: center;
   width: 100vw;
   height: 100vh;
-  padding: 30px;
+  padding: 20px;
 }
 .form {
   flex: 1;
@@ -19,18 +19,4 @@
   padding: 5rem ;
   border-radius: 1rem ;
   box-shadow: -3px -2px 20px 1px #202020;
-}
-.incitingText {
-  color: #1900d5;
-  margin-left: 10px;
-  cursor: pointer;
-}
-
-.incitingButton{
-  border-radius: 4px;
-  background-color: #1900d5 !important;
-  box-shadow: 0px 1px 7px #332a76 !important;
-  margin-top: 2rem;
-  margin-bottom: 2rem;
-  width: 60% ;  
 }
