+++ src/pages/home/index.ts
@@ -1,0 +1,3 @@
+export { HomePage } from "./HomePage";
+
+
