+++ src/pages/home/styles.module.css
@@ -45,13 +45,13 @@
 }
 
 .lineChartBlock {
-  padding: 16px 20px;
+  padding: 20px 16px;
   margin-bottom: 32px;
   min-height: 270px;
   border-radius: 12px;
 }
 .topPriceRoot {
-  padding: 16px 20px;
+  padding: 20px 16px;
   margin-bottom: 32px;
   min-height: 270px;
   border-radius: 12px;
