+++ src/pages/news/NewsPage.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+export const NewsPage = () => {
+  return (
+    <div>
+      <h1>News page</h1>
+    </div>
+  );
+};
+export default NewsPage;
