--- src/pages/news/index.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const NewsComponent = () => {
-  return (
-    <div><h1>News page</h1></div>
-  )
-}
-
-export default NewsComponent
