+++ src/pages/settings/SettingsPage.tsx
@@ -1,11 +1,10 @@
 import React from "react";
 
-const SettingsComponent = () => {
+export const SettingsPage = () => {
   return (
     <div>
       <h1>Settings page</h1>
     </div>
   );
 };
-
-export default SettingsComponent;
+export default SettingsPage;
