+++ src/pages/settings/index.ts
@@ -1,0 +1,3 @@
+export { SettingsPage } from "./SettingsPage";
+
+
