+++ src/pages/settings/newFile.tsx
@@ -1,0 +1,3 @@
+import { SettingsPage } from "./SettingsPage";
+
+export default SettingsPage;
