+++ src/pages/single-asset/SingleAssetPage.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+import { useNavigate } from "react-router-dom";
+
+export const SingleAssetPage = () => {
+  const navigate = useNavigate()
+  return (
+    <div>
+      <h1 onClick={() => navigate(-1)}>Go back</h1>
+    </div>
+  );
+};
+
+export default SingleAssetPage;
