+++ src/pages/single-asset/index.ts
@@ -1,0 +1,3 @@
+export { SingleAssetPage } from "./SingleAssetPage";
+
+
