+++ src/pages/watchlist/WatchlistPage.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+export const WatchlistPage = () => {
+  return (
+    <div>
+      <h1>Watchlist</h1>
+    </div>
+  );
+};
+export default WatchlistPage;
