+++ src/pages/watchlist/index.ts
@@ -1,0 +1,3 @@
+export { WatchlistPage } from "./WatchlistPage";
+
+
