--- src/pages/watchlist/index.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const WatchlistComponent = () => {
-  return (
-    <div><h1>Watchlist</h1></div>
-  )
-}
-
-export default WatchlistComponent
